requirements.txt
python-telegram-bot[job-queue]==21.6
google-generativeai==0.8.3
sqlalchemy==2.0.35
matplotlib==3.9.2
pandas==2.2.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9
